Updates fonts.
Replaces the font combination 'Coming Soon' and 'Lato' with the
more elegant 'Spinnaker' and 'EB Garamond'. These fonts represent
fewer kbs, improving performance. They are also more mature and
flexible than the former fonts.

This is part of an overdue design decision to move away from the
"hand-drawn" look to a more elegant, magazine-like appearance that
appeals to a wider audience.